+++ src/translations/fr.ts
@@ -8,6 +8,9 @@ export const fr: Translations = {
   grid_coupling: {
     title_default: 'Couplage réseau',
   },
+  pv_system: {
+    title_default: 'Production photovoltaïque',
+  },
   confirm: {
     toggle: 'Êtes-vous sûr de vouloir commuter « {name} » ?',
   },